cmake: fix when cross compiling
In order to generate protobuf files CMake need to use the protoc
and grpc-cpp-plugin compiled for the host architecture.

Unfortunately, the protoc and grpc-cpp-plugin in the gRPC CMake
configuration file are the one for the target architecture.

Fix this by properly finding the correct executable when
CMake is cross compiling.

Signed-off-by: Clément Péron <peron.clem@gmail.com>